+++ src/meeting.ts
@@ -9,7 +9,7 @@ import { ParticipantModel } from "./Models/participant-model";
 import { ConusmaException } from "./Exceptions/conusma-exception";
 import { ConusmaWorker } from "./conusma-worker";
 import InCallManager from 'react-native-incall-manager';
-import DeviceInfo from 'react-native-device-info';
+
 import { MediaServer } from "./media-server";
 import { Connection } from "./connection";
 
@@ -27,7 +27,7 @@ export class Meeting {
     private appService: AppService;
 
     public isReceviedClose: boolean = false;
-    private cameraCrashCounter = 2;
+    
 
     constructor(ownerUser: MeetingUserModel, appService: AppService) {
         registerGlobals();
@@ -91,9 +91,6 @@ export class Meeting {
         this.observers.forEach(observer => observer());
     }
     
-
-    
-
     private async getMediaServer(meetingUserId: string) {
         return await this.appService.getMediaServer(meetingUserId);
     }
@@ -140,71 +137,14 @@ export class Meeting {
         }
 
     }
-    public switchCamera() {
-        try {
-            if (this.mediaServerClient != null && this.mediaServerClient.Stream != null) {
-
-                var deviceModel: string = DeviceInfo.getModel();
-                deviceModel = deviceModel.toLowerCase();
-                if (deviceModel.includes('sm-n975') || deviceModel.includes('sm-g981') || deviceModel.includes('sm-g980')) {
-                    if (this.cameraCrashCounter <= 0) {
-                        throw new Error("camera switching is not supported on this model ");
-                    }
-                }
-                this.mediaServerClient.Stream.getVideoTracks()[0]._switchCamera();
-                this.cameraCrashCounter--;
-                return this.mediaServerClient.Stream;
-            }
-            else {
-                throw new Error("stream not found, first call enableAudioVideo function");
-            }
-
-        } catch (error) {
-            throw new ConusmaException("switchCamera", "camera switching failed, please check detail exception", error);
-        }
-    }
-    public toggleAudio() {
-        try {
-            if (this.mediaServerClient != null && this.mediaServerClient.Stream != null) {
-                this.mediaServerClient.Stream.getTracks().forEach((t: any) => {
-                    if (t.kind === 'audio') {
-                        t.enabled = !t.enabled;
-                        this.isAudioActive = t.enabled;
-
-                    }
-                });
-                return <MediaStream>this.mediaServerClient.Stream;
-            }
-            else {
-                throw new ConusmaException("toggleAudio", "stream not found, first call enableAudioVideo function");
-            }
-
-        } catch (error) {
-            throw new ConusmaException("toggleAudio", "toggleAudio failed", error);
-        }
-    }
-    public toggleVideo() {
-        try {
-            if (this.mediaServerClient != null && this.mediaServerClient.Stream != null) {
-                this.isVideoActive = !this.isVideoActive;
-                this.mediaServerClient.Stream.getVideoTracks()[0].enabled = this.isVideoActive;
-                return <MediaStream>this.mediaServerClient.Stream;
-            }
-            else {
-                throw new ConusmaException("toggleVideo", "stream not found, first call enableAudioVideo function");
-            }
-
-        } catch (error) {
-            throw new ConusmaException("toggleVideo", "toggleVideo failed", error);
-        }
-    }
+    
+    
 
     public async enableAudioVideo() {
         try {
 
         } catch (error) {
             throw new ConusmaException("enableAudioVideo", "can not read stream , please check exception ", error);
-
         }
         const isFrontCamera = true;
         const devices = await mediaDevices.enumerateDevices();
@@ -212,10 +152,6 @@ export class Meeting {
         const videoSourceId = devices.find(
             (device: any) => device.kind === 'videoinput' && device.facing === facing,
         );
-        if (videoSourceId) {
-            this.hasCamera = true;
-            this.hasMicrophone = true; // TODO: Check audio source first
-        }
         const facingMode = isFrontCamera ? 'user' : 'environment';
         const constraints: any = {
             audio: true,
@@ -230,13 +166,11 @@ export class Meeting {
             },
         };
         const newStream: MediaStream = await mediaDevices.getUserMedia(constraints);
-        this.isAudioActive = true;
-        this.isVideoActive = true;
         return newStream;
     }
     public async connectMeeting() {
         try {
-            await this.appService.connectMeeting(this.meetingUser);
+            await this.appService.connectMeeting(this.ownerUser);
             console.log("User connected to the meeting.");
         } catch (error) {
             throw new ConusmaException("connectMeeting", "can not connect meeting , please check exception", error);
@@ -245,58 +179,13 @@ export class Meeting {
     }
     public async isApproved() {
         try {
-            return await this.appService.isApproved(this.meetingUser.Id);
+            return await this.appService.isApproved(this.ownerUser.Id);
 
         } catch (error) {
-            throw new ConusmaException("isApproved", "user is not approved , please check exception ", error);
+            throw new ConusmaException("isApproved", "user is not approved, please check exception ", error);
         }
     }
-    public async consume(producerUser: MeetingUserModel) {
-        try {
-            var result = await this.createConsumerTransport(producerUser);
-            this.consumerTransports.push(result);
-            return <MediaStream>result.RemoteStream;
-        } catch (error) {
-
-            throw new ConusmaException("consume", producerUser.Id + "The stream of the user is currently not captured. User connection information is out of date.", error);
-        }
-    }
-
-    public async closeConsumer(user: MeetingUserModel) {
-        try {
-            var index = 0;
-            for (let item of this.consumerTransports) {
-                if (item.MeetingUserId == user.Id) {
-                    if (item.transport) {
-                        item.transport.close();
-                    }
-                    break;
-                }
-                index++;
-            };
-            this.removeItemOnce(this.consumerTransports, index);
-        } catch (error) {
-            throw new ConusmaException("isApproved", "user is not approved , please check exception ", error);
-        }
-
-    }
-
-    private removeItemOnce(arr: any, index: any) {
-        if (index > -1) {
-            arr.splice(index, 1);
-        }
-        return arr;
-    }
-
-    public setSpeaker(enable: boolean) {
-        try {
-            InCallManager.setSpeakerphoneOn(enable);
-            InCallManager.setForceSpeakerphoneOn(enable);
-
-        } catch (error) {
-            throw new ConusmaException("setSpeaker", "setSpeaker undefined error", error);
-        }
-    }
+    
     private waitWhoAreYou(socket: any) {
         return new Promise(resolve => {
             socket.on("WhoAreYou")
@@ -306,155 +195,10 @@ export class Meeting {
             }
         });
     }
-    private async createConsumerTransport(user: MeetingUserModel) {
-        var targetMediaServerClient: MediaServer = <MediaServer>this.mediaServerList.find((ms: any) => ms.Id == user.MediaServerId);
-
-        if (targetMediaServerClient == null) {
-            targetMediaServerClient = new MediaServer();
-
-            var mediaServerInfo: any = await this.appService.getMediaServerById(this.meetingUser.Id, user.MediaServerId);
-            if (mediaServerInfo == null) {
-                throw new ConusmaException("createConsumerTransport", "Media server not found. (Id: " + user.MediaServerId + ")");
-            }
-            targetMediaServerClient.Id = mediaServerInfo.Id;
-            targetMediaServerClient.socket = io.connect(mediaServerInfo.ConnectionDnsAddress + ":" + mediaServerInfo.Port);
-
-            /*console.log("waiting WhoAreYou signal...");
-            var waitResponse = await this.waitWhoAreYou(targetMediaServerClient.socket);
-            console.log("WhoAreYou signal came.");*/
-
-            var userInfoData = { 'MeetingUserId': this.meetingUser.Id, 'Token': this.appService.getJwtToken() };
-            let setUserInfo = await this.signal('UserInfo', userInfoData, targetMediaServerClient.socket);
-            console.log("UserInfo signal came.");
-
-            let routerRtpCapabilities = await this.signal('getRouterRtpCapabilities', null, targetMediaServerClient.socket);
-            console.log("routerRtpCapabilities " + JSON.stringify(routerRtpCapabilities));
-
-            const handlerName = mediaServerClient.detectDevice();
-            if (handlerName) {
-                console.log("detected handler: %s", handlerName);
-            } else {
-                console.error("no suitable handler found for current device");
-            }
-
-            targetMediaServerClient.mediaServerDevice = new mediaServerClient.Device({
-                handlerName: handlerName
-            });
-            console.log("mediaServerDevice loading...");
-            await targetMediaServerClient.mediaServerDevice.load({ routerRtpCapabilities });
-            console.log("mediaServerDevice loaded.");
-            this.mediaServerList.push(targetMediaServerClient);
-            return await this.createConsumerChildFunction(targetMediaServerClient, user);
-        } else {
-            return await this.createConsumerChildFunction(targetMediaServerClient, user);
-        }
-    }
-
-    private async createConsumerChildFunction(targetMediaServerClient: MediaServer, user: MeetingUserModel) {
-        if (targetMediaServerClient != null && targetMediaServerClient.socket != null) {
-            console.log("createConsumerChildFunction start.");
-
-            var consumerTransport: any = new Object();
-            consumerTransport.MediaServer = targetMediaServerClient;
-            consumerTransport.MeetingUserId = user.Id;
-            var transportOptions = await this.signal("createConsumerTransport", { MeetingUserId: user.Id }, targetMediaServerClient.socket);
-            consumerTransport.MediaServerSocketId = user.MediaServerSocketId;
-            consumerTransport.transportId = transportOptions.Id;
-            consumerTransport.transport = await targetMediaServerClient.mediaServerDevice.createRecvTransport(transportOptions.transportOptions);
-            consumerTransport.transport.on("connect", async ({ dtlsParameters }: any, callback: any, errback: any) => {
-                this.signal("connectConsumerTransport", { consumerTransportId: consumerTransport.transportId, dtlsParameters: dtlsParameters }, targetMediaServerClient.socket)
-                    .then(callback)
-                    .catch(errback);
-            });
-            consumerTransport.RemoteStream = new MediaStream();
-            consumerTransport.Camera = user.Camera;
-            consumerTransport.Mic = user.Mic;
-            consumerTransport.ShareScreen = user.ShareScreen;
-            console.log("createConsumerChildFunction creating the consumer.");
-
-            if (user.Camera || user.ShareScreen) {
-                await this.addConsumer(consumerTransport, "video");
-            }
-
-            if (user.Mic) {
-                await this.addConsumer(consumerTransport, "audio");
-            }
-            return consumerTransport;
-        } else {
-            throw new ConusmaException("createConsumerChildFunction", "No socket connection.");
-        }
-    }
-
-    private async addConsumer(consumerTransport: any, kind: string) {
-        if (consumerTransport != null) {
-            if (kind == "video") {
-                consumerTransport.videoConsumer = await this.consumeTransport(consumerTransport, "video");
-                this.resumeConsumer(consumerTransport, "video");
-                consumerTransport.RemoteStream.addTrack(consumerTransport.videoConsumer.track);
-            } else {
-                consumerTransport.audioConsumer = await this.consumeTransport(consumerTransport, "audio");
-                this.resumeConsumer(consumerTransport, "audio");
-                consumerTransport.RemoteStream.addTrack(consumerTransport.audioConsumer.track);
-                consumerTransport.audioConsumer.resume();
-            }
-        }
-    }
-
-    private async consumeTransport(consumerTransport: any, trackKind: string) {
-        const { rtpCapabilities } = consumerTransport.MediaServer.mediaServerDevice;
-        const data = await this.signal("consume", { consumerTransportId: consumerTransport.transportId, rtpCapabilities: rtpCapabilities, kind: trackKind }, consumerTransport.MediaServer.socket)
-            .catch(err => {
-                throw new ConusmaException("consumeTransport", "Consume error.", err);
-            });
-        const {
-            producerId,
-            id,
-            kind,
-            rtpParameters,
-        } = data;
-        let codecOptions = {};
-        const consumer = await consumerTransport.transport.consume({
-            id,
-            producerId,
-            kind,
-            rtpParameters,
-            codecOptions,
-        });
-        return consumer;
-    }
-
-    private async resumeConsumer(consumerTransport: any, kind: string) {
-        this.signal('resume', { consumerTransportId: consumerTransport.transportId, kind: kind }, consumerTransport.MediaServer.socket);
-    }
-    private async pauseConsumer(consumerTransport: any, kind: string) {
-        try {
-            if (consumerTransport != null && consumerTransport.videoConsumer != null) {
-                if (kind == 'video') {
-                    await this.signal('pause', {
-                        kind: 'video',
-                        consumerTransportId: consumerTransport.transportId
-                    }, consumerTransport.MediaServer.socket);
-                    await consumerTransport.videoConsumer.pause();;
-                    consumerTransport.RemoteStream.removeTrack(consumerTransport.videoConsumer.track);
-                }
-                else if (kind == 'audio' && consumerTransport.audioConsumer != null) {
-                    await this.signal('pause', {
-                        kind: 'audio',
-                        consumerTransportId: consumerTransport.transportId
-                    }, consumerTransport.MediaServer.socket);
-                    await consumerTransport.audioConsumer.pause();
-                    consumerTransport.RemoteStream.removeTrack(consumerTransport.audioConsumer.track);
-                }
-            }
-        } catch (error) {
-
-        }
-    }
-
     public async getAllUsers() {
         try {
-            if (this.meetingUser != null) {
-                return <MeetingUserModel[]>await this.appService.getMeetingUserList({ 'MeetingUserId': this.meetingUser.Id });
+            if (this.ownerUser != null) {
+                return <MeetingUserModel[]>await this.appService.getMeetingUserList({ 'MeetingUserId': this.ownerUser.Id });
             } else {
                 return [];
             }
@@ -467,8 +211,8 @@ export class Meeting {
 
     public async getProducerUsers() {
         try {
-            if (this.meetingUser != null) {
-                var users = await this.appService.getMeetingUserList({ 'MeetingUserId': this.meetingUser.Id });
+            if (this.ownerUser != null) {
+                var users = await this.appService.getMeetingUserList({ 'MeetingUserId': this.ownerUser.Id });
                 var result: MeetingUserModel[] = [];
                 users.forEach((item: any) => {
                     if (item.Camera == true) {
